+++ filetype_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/h2non/filetype/types"
+	"github.com/miihael/filetype/types"
 )
 
 func TestConcurrent(t *testing.T) {